bi-dev-master/requirements.txt
alembic==0.9.8
amqp==2.2.2
asn1crypto==0.24.0
Babel==2.5.3
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.3
boto3==1.5.34
botocore==1.8.48
celery==4.1.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==1.9
docutils==0.14
enum34==1.1.6
Flask==0.12.2
Flask-AppBuilder==1.9.4
Flask-Babel==0.11.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-Migrate==2.0.3
Flask-OpenID==1.2.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.2
Flask-WTF==0.14.2
flower==0.9.1
future==0.16.0
futures==3.2.0
gunicorn==19.7.1
humanize==0.5.1
idna==2.5
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
MySQL-python==1.2.5
mysqlclient==1.3.12
numpy==1.14.1
pandas==0.20.3
parsedatetime==2.0
pathlib2==2.3.0
pycparser==2.18
pydruid==0.3.1
PyHive==0.5.1
PyMySQL==0.8.0
python-dateutil==2.6.0
python-editor==1.0.3
python-openid==2.2.5
pytz==2018.3
PyYAML==3.12
requests==2.17.3
s3transfer==0.1.13
sasl==0.2.1
scandir==1.7
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.9
SQLAlchemy-Utils==0.32.16
sqlparse==0.2.3
thrift==0.11.0
thrift-sasl==0.3.0
tornado==4.2
Unidecode==1.0.22
urllib3==1.21.1
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.1
